python/requirements.txt
azure-identity>=1.17.0
azure-mgmt-cosmosdb>=9.0.0
requests>=2.31.0
